refacor: package restructure and package export updates
