+++ package.json
@@ -5,15 +5,13 @@
 	"exports": {
 		"./components": {
 			"types": "./dist/components/index.d.ts",
+			"require": "./dist/components.cjs",
 			"import": "./dist/components.js"
 		},
 		"./core": {
 			"types": "./dist/core/index.d.ts",
+			"require": "./dist/core.cjs",
 			"import": "./dist/core.js"
-		},
-		"./services": {
-			"types": "./dist/services/index.d.ts",
-			"import": "./dist/services.js"
 		},
 		"./lib.css": {
 			"import": "./dist/lib.css",
