+++ src/components/bo-button/bo-button.test.ts
@@ -5,10 +5,10 @@ import BoIcon from '@/components/bo-icon/bo-icon.vue';
 import BoLoadingRing from '@/components/bo-loading-ring/bo-loading-ring.vue';
 import BoLoadingSpinner from '@/components/bo-loading-spinner/bo-loading-spinner.vue';
 import { BoText } from '@/components/bo-text';
-import { AriaLive } from '@/core/accessibility';
-import { BoLoaderType } from '@/core/loader';
-import { BoSize } from '@/core/size';
-import { BoVariant } from '@/core/variant';
+import { AriaLive } from '@/core/constants/accessibility';
+import { BoLoaderType } from '@/core/constants/loader';
+import { BoSize } from '@/core/constants/size';
+import { BoVariant } from '@/core/constants/variant';
 import { BoButtonType } from './bo-button';
 import { Icon } from '@/components/bo-icon/bo-icon';
