+++ src/components/bo-button/bo-button.vue
@@ -61,13 +61,13 @@
 	import BoLoadingSpinner from '@/components/bo-loading-spinner/bo-loading-spinner.vue';
 	import { BoFontSize, BoFontWeight, BoTextVariant } from '@/components/bo-text/bo-text.js';
 	import boText from '@/components/bo-text/bo-text.vue';
-	import { AriaLive } from '@/core/accessibility.js';
-	import type { ConditionalCssProperties } from '@/core/css';
-	import { BoLoaderType } from '@/core/loader.js';
-	import { BoSize } from '@/core/size.js';
-	import { BoVariant } from '@/core/variant.js';
-	import { ColorService } from '@/services/color-service.js';
-	import { IdentityService } from '@/services/identity-service.js';
+	import { AriaLive } from '@/core/constants/accessibility.js';
+	import type { ConditionalCssProperties } from '@/core/constants/css.js';
+	import { BoLoaderType } from '@/core/constants/loader.js';
+	import { BoSize } from '@/core/constants/size.js';
+	import { BoVariant } from '@/core/constants/variant.js';
+	import { ColorService } from '@/core/services/color-service.js';
+	import { IdentityService } from '@/core/services/identity-service.js';
 	import { computed, type HTMLAttributes, type StyleValue } from 'vue';
 	import { BoButtonType, type BoButtonProps } from './bo-button.js';
