+++ src/core/constants/svg.ts
