+++ src/core/constants/variant.ts
