+++ src/core/index.ts
@@ -1,7 +1,15 @@
-export * from './accessibility.js';
-export * from './loader.js';
-export * from './size.js';
-export * from './variant.js';
-export * from './local-storage.js';
-export * from './css.js';
-export * from './svg.js';
+/** Constants */
+
+export * from './constants/accessibility.js';
+export * from './constants/loader.js';
+export * from './constants/size.js';
+export * from './constants/variant.js';
+export * from './constants/local-storage.js';
+export * from './constants/css.js';
+export * from './constants/svg.js';
+
+/** Services */
+
+export * from './services/color-service.js';
+export * from './services/identity-service.js';
+export * from './services/theme-service.js';
