+++ src/core/services/color-service.ts
