--- src/services/index.ts
@@ -1,3 +1,0 @@
-export * from './color-service.js';
-export * from './identity-service.js';
-export * from './theme-service.js';
